pyproject.toml
[project]
name = "cloud-downscaling"
version = "0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "einops>=0.8.1",
    "h5py>=3.15.1",
    "lightning>=2.5.5",
    "numpy>=2.1.2",
    "python-dotenv>=1.2.1",
    "timm>=1.0.19",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]

[tool.setuptools.packages.find]
where = ["src"]
